io_quizzes_create/requirements.txt
Flask==1.1.1
get-docker-secret==1.0.1
psycopg2-binary==2.8.5
